requirements.txt
aiogram>=3.13.0,<4.0.0
aiosqlite>=0.19.0
APScheduler>=3.10.4
python-dotenv>=1.0.0